@@ -13,16 +13,16 @@
   "dependencies": {
     "mapbox-gl": "^0.49.0-beta.1",
     "normalize.css": "^8.0.1",
-    "react": "^16.3.1",
-    "react-dom": "^16.3.1",
-    "react-modal": "^3.4.4",
-    "react-redux": "^5.0.7",
-    "react-router-dom": "^4.2.2",
-    "react-transition-group": "^2.3.1",
-    "redux": "^4.0.0",
-    "redux-thunk": "^2.2.0",
-    "simple-oauth2": "^2.2.1",
-    "styled-components": "^3.4.5"
+    "randombytes": "^2.1.0",
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0",
+    "react-modal": "^3.11.1",
+    "react-redux": "^7.1.3",
+    "react-router-dom": "^5.1.2",
+    "react-transition-group": "^4.3.0",
+    "redux": "^4.0.4",
+    "redux-thunk": "^2.3.0",
+    "styled-components": "^4.4.1"
   },
   "devDependencies": {
     "babel-core": "^6.26.3",
@@ -38,11 +38,12 @@
     "file-loader": "^1.1.11",
     "html-webpack-plugin": "^3.2.0",
     "node-sass": "^4.8.3",
-    "oidc-client": "^1.5.2",
+    "oidc-client": "^1.9.1",
     "sass-loader": "^6.0.7",
+    "simple-oauth2": "^3.1.0",
     "style-loader": "^0.20.3",
     "webpack": "^4.5.0",
-    "webpack-cli": "^2.0.14",
+    "webpack-cli": "^3.1.1",
     "webpack-dev-server": "^3.1.3"
   }
 }
